Move loadResourceAsString to Resource class
It doesn't make much sense for this function to sit
inside IconResources, as it isn't an IconResource.
